+++ ms_invoice_management/infrastructure/driven-adapters/jpa-repository/src/main/java/co/com/management/jpa/persistence/invoice/InvoiceDao.java
@@ -1,27 +1,19 @@
 package co.com.management.jpa.persistence.invoice;
 
-import co.com.management.jpa.persistence.client.ClientDao;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
-import jakarta.persistence.PrePersist;
-import jakarta.persistence.PreUpdate;
 import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.time.LocalDate;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.UUID;
 
 @Getter
 @Setter
@@ -35,7 +27,7 @@ public class InvoiceDao {
     private String id;
 
     @Column(name="total_amount",nullable = false)
-    private Double totalAmount;
+    private BigDecimal totalAmount;
 
     @OneToMany(mappedBy = "invoice", fetch = FetchType.EAGER, cascade = CascadeType.ALL)
     List<ProductDao> products;
